[#61] Refactor : remove unused class
