+++ src/main/java/com/m9d/sroom/config/AuthInterceptor.java
@@ -9,16 +9,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.codec.binary.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StreamUtils;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.util.Enumeration;
 import java.util.Map;
 
 @Component
@@ -26,17 +21,11 @@ import java.util.Map;
 public class AuthInterceptor implements HandlerInterceptor {
 
     @Autowired
-    public AuthInterceptor() {
-
-    }
-
-    @Autowired
     private JwtUtil jwtUtil;
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         log.info("Request URI: {}", request.getRequestURI());
-        Enumeration<String> headerNames = request.getHeaderNames();
         log.debug("HTTP Method: {}", request.getMethod());
 
         if (!(handler instanceof HandlerMethod))
