+++ src/main/java/com/m9d/sroom/dashboard/dto/response/Dashboard.java
@@ -1,4 +1,4 @@
-package com.m9d.sroom.dashbord.dto.response;
+package com.m9d.sroom.dashboard.dto.response;
 
 import com.m9d.sroom.course.dto.response.CourseInfo;
 import io.swagger.v3.oas.annotations.media.Schema;
