+++ src/main/java/com/m9d/sroom/global/error/NotMatchException.java
@@ -1,15 +1,16 @@
-package com.m9d.sroom.config.error;
+package com.m9d.sroom.global.error;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
+
 @Getter
-public abstract class DuplicationException extends RuntimeException {
+public abstract class NotMatchException extends RuntimeException {
 
     private final int statusCode = HttpStatus.NOT_FOUND.value();
     private String message;
 
-    public DuplicationException(String message) {
+    public NotMatchException(String message) {
         this.message = message;
     }
 }
