+++ src/main/java/com/m9d/sroom/lecture/exception/LectureNotFoundException.java
@@ -1,6 +1,6 @@
 package com.m9d.sroom.lecture.exception;
 
-import com.m9d.sroom.config.error.NotFoundException;
+import com.m9d.sroom.global.error.NotFoundException;
 
 public class LectureNotFoundException extends NotFoundException {
