+++ src/main/java/com/m9d/sroom/member/exception/MemberNotMatchException.java
@@ -1,6 +1,6 @@
 package com.m9d.sroom.member.exception;
 
-import com.m9d.sroom.config.error.NotMatchException;
+import com.m9d.sroom.global.error.NotMatchException;
 
 public class MemberNotMatchException extends NotMatchException {
