+++ src/test/java/com/m9d/sroom/dashboard/service/DashboardServiceTest.java
@@ -1,7 +1,7 @@
 package com.m9d.sroom.dashboard.service;
 
 
-import com.m9d.sroom.dashbord.dto.response.Dashboard;
+import com.m9d.sroom.dashboard.dto.response.Dashboard;
 import com.m9d.sroom.member.domain.Member;
 import com.m9d.sroom.util.ServiceTest;
 import org.junit.jupiter.api.Assertions;
