+++ src/test/java/com/m9d/sroom/util/SroomTest.java
@@ -4,8 +4,8 @@ package com.m9d.sroom.util;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.m9d.sroom.course.repository.CourseRepository;
 import com.m9d.sroom.course.service.CourseService;
-import com.m9d.sroom.dashbord.repository.DashboardRepository;
-import com.m9d.sroom.dashbord.service.DashboardService;
+import com.m9d.sroom.dashboard.repository.DashboardRepository;
+import com.m9d.sroom.dashboard.service.DashboardService;
 import com.m9d.sroom.lecture.service.LectureService;
 import com.m9d.sroom.member.repository.MemberRepository;
 import com.m9d.sroom.member.service.MemberService;
